Add skeleton code for adding destination.
Adding a destination point works locally. It doesn't do anything, but you can place it with a long press. Still need to communicate it over the server and do some processing to figure out when you reach it. David should be able to build out the atSameLocation function in MapboxMap.js independently.